Replaced named pipe communication in VFD SWMR test programs with socket-based communication instead. Implemented new scripts to test socket functionality locally and over networked environment.
- Introduced a new script `exec_nfs_socket_test.sh` to facilitate testing over NFS mounted directories using socket communication.
- The script allows configuration of various test parameters, including test type, role (reader/writer), and optional metadata directory.
- Implemented functions to handle test execution, signaling between reader and writer processes, and error handling for NFS I/O issues.
- Supports multiple test scenarios including 'attrdset', 'bigset', 'dsetops', 'dsetchks', 'gfail', 'group_basic', 'group_attrs', 'os_group_attrs', and 'zoo'.
- Configured command paths for writer and reader executables based on the selected test.

TO REPRODUCE ERRORS WITH SOCKET CODE.
First, make sure that you are running the scripts from within the same NFS mounted directory on two separate devices. One script gets the 'w' role, and the other gets the 'r' role.
Run the 'attrdset' or 'group_basic' tests to see if they work or not. These two tests are pretty small, so you shouldnt need to edit the option sets to get to the errors.
To see the errors with dsetchks and dsetops, you can just can either run the 'dsetchks' and 'dsetops' tests as they are, or you can go to lines 215 and 250 to comment out every option in the list except for the ones that have the # FAILS comment next to them.
For the bigset test, you can edit the vfd_swmr_bigset_writer.c code and comment out the #define ADD_UNIQUE_STEP_FILE 1 line (or undef ADD_UNIQUE_STEP_FILE after), rebuild, and then test the bigset test over NFS to see the failure that happens there.